Make list of match map filters dynamic by mode.
Use the list of maps by mode endpoint to reduce all maps
from seasons by game mode into a set. There are a few maps
that have duplicate translations, so special handling will
be needed to address duplicate labels in the dropdown. There
is at least one map which does not have a translation and also
has an invalid character, "." in the name which will have to
accounted for when retrieving the translation - "lastrefuge.anon".