Add css variables for theming.
Vuetify will dynamically restyle the theme on dark/light mode
switch. To take advantage of that, we override some of the variables
that are used by the theme, and reference them to CSS variables that
we can set in the specific race scss partials, or just by overriding
them at theme switch time. Both approaches are taken here, one for
doing the glass, and the other for setting primary colors and
button foreground text.