Add map filter to matches. (#309)
* Add map filter to matches.

The match select uses a hard-coded list of maps. This might be
more appropriate to get from a API. I'm not 100% sure that the list
hard-coded in this commit is the most correct list of maps, but it
seems to be pretty close.

* Make list of match map filters dynamic by mode.

Use the list of maps by mode endpoint to reduce all maps
from seasons by game mode into a set. There are a few maps
that have duplicate translations, so special handling will
be needed to address duplicate labels in the dropdown. There
is at least one map which does not have a translation and also
has an invalid character, "." in the name which will have to
accounted for when retrieving the translation - "lastrefuge.anon".

Co-authored-by: s.heiss <simonheiss87@gmail.com>