WYSIWYG for admin news and simpler race theming. (#351)
* Add tiptap wysiwyg to admin page.

Version 2 of tiptap will provide proper typings for the components.
For now, relying no some community submitted typing. Still need
to explore binding the wysiwyg content to the component/store
and handling image uploads to a CDN.

* Add css variables for theming.

Vuetify will dynamically restyle the theme on dark/light mode
switch. To take advantage of that, we override some of the variables
that are used by the theme, and reference them to CSS variables that
we can set in the specific race scss partials, or just by overriding
them at theme switch time. Both approaches are taken here, one for
doing the glass, and the other for setting primary colors and
button foreground text.

* Remove commented lines from undead styles, remove fibers.

Co-authored-by: Aaron Jewell <ajewell@getacceptd.com>